refactor(test): extract shared TestKitMetrics from TestStream + CrashRestartHarness
TestStream (live drive) and CrashRestartHarness (crash/restart) hand-rolled
byte-identical quiescence logic: the same 4 metric-key constants
(messagesReceived/Processed/Errors/inFlight), the same accountedFor formula,
and a near-identical quiescent() check (TestStream on the instance's
consumer, CrashRestartHarness static). That is the 'parallel lists that
drift' hazard — a metric-key rename or an accounting-formula change had to be
made in two places or the two harnesses would silently disagree.

Extract a package-private TestKitMetrics holding the constants + accountedFor
(now private) + quiescent(KPipeConsumer, long). Both harnesses call
TestKitMetrics.quiescent(consumer, target); each keeps its own deadline loop.
One source of truth, unit-testable in isolation. CLOSE_AGE_CAP stays on
TestStream — it's a batch-policy age cap, not a metrics concern.

Behavior-preserving: :lib:kpipe-test:test green.
